refactor jmri.Section and jmri.SectionManager
