+++ java/src/jmri/SectionManager.java
@@ -7,12 +7,9 @@ import java.util.Set;
 
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
-import jmri.jmrit.display.layoutEditor.LayoutEditor;
 import jmri.managers.AbstractManager;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import jmri.jmrit.display.layoutEditor.LayoutBlockManager;
-import jmri.jmrit.display.layoutEditor.LayoutBlock;
+
+import jmri.jmrit.display.layoutEditor.LayoutEditor;
 
 /**
  * Basic Implementation of a SectionManager.
@@ -37,32 +34,9 @@ import jmri.jmrit.display.layoutEditor.LayoutBlock;
  *
  * @author Dave Duchamp Copyright (C) 2008
  */
-public class SectionManager extends AbstractManager<Section> implements InstanceManagerAutoDefault {
+public interface SectionManager extends Manager<Section> {
 
-    public SectionManager() {
-        super();
-        addListeners();
-    }
-    
-    final void addListeners() {
-        InstanceManager.getDefault(SensorManager.class).addVetoableChangeListener(this);
-        InstanceManager.getDefault(BlockManager.class).addVetoableChangeListener(this);
-    }
-
-    @Override
-    public int getXMLOrder() {
-        return Manager.SECTIONS;
-    }
-
-    @Override
-    public char typeLetter() {
-        return 'Y';
-    }
-
-    @Override
-    public Class<Section> getNamedBeanClass() {
-        return Section.class;
-    }
+    // void addListeners();
 
     /**
      * Create a new Section if the Section does not exist.
@@ -75,40 +49,8 @@ public class SectionManager extends AbstractManager<Section> implements Instance
      *         Section.
      */
     @Nonnull
-    public Section createNewSection(@Nonnull String systemName, String userName) throws IllegalArgumentException {
-        Objects.requireNonNull(systemName, "SystemName cannot be null. UserName was " + ((userName == null) ? "null" : userName));  // NOI18N
-        // check system name
-        if (systemName.isEmpty()) {
-            throw new IllegalArgumentException("Section System Name Empty");
-            // no valid system name entered, return without creating
-        }
-        String sysName = systemName;
-        if (!sysName.startsWith(getSystemNamePrefix())) {
-            sysName = makeSystemName(sysName);
-        }
-        // Check that Section does not already exist
-        Section y;
-        if (userName != null && !userName.isEmpty()) {
-            y = getByUserName(userName);
-            if (y != null) {
-                throw new IllegalArgumentException("Section Already Exists with UserName " + userName);
-            }
-        }
-        y = getBySystemName(sysName);
-        if (y != null) {
-            throw new IllegalArgumentException("Section Already Exists with SystemName " + sysName);
-        }
-        // Section does not exist, create a new Section
-        y = new Section(sysName, userName);
-        // save in the maps
-        register(y);
+    public Section createNewSection(@Nonnull String systemName, String userName) throws IllegalArgumentException;
 
-        // Keep track of the last created auto system name
-        updateAutoNumber(systemName);
-
-        return y;
-    }
-    
     /**
      * Create a New Section with Auto System Name.
      * @param userName UserName for new Section
@@ -117,20 +59,14 @@ public class SectionManager extends AbstractManager<Section> implements Instance
      *          unable to create a new Section.
      */
     @Nonnull
-    public Section createNewSection(String userName) throws IllegalArgumentException {
-        return createNewSection(getAutoSystemName(), userName);
-    }
+    public Section createNewSection(String userName) throws IllegalArgumentException;
 
     /**
      * Remove an existing Section.
      *
      * @param y the section to remove
      */
-    public void deleteSection(Section y) {
-        // delete the Section
-        deregister(y);
-        y.dispose();
-    }
+    public void deleteSection(Section y);
 
     /**
      * Get an existing Section. First look up assuming that name is a User
@@ -140,13 +76,7 @@ public class SectionManager extends AbstractManager<Section> implements Instance
      *             followed by system names
      * @return the found section of null if no matching Section found
      */
-    public Section getSection(String name) {
-        Section y = getByUserName(name);
-        if (y != null) {
-            return y;
-        }
-        return getBySystemName(name);
-    }
+    public Section getSection(String name);
 
     /**
      * Validate all Sections.
@@ -156,26 +86,7 @@ public class SectionManager extends AbstractManager<Section> implements Instance
      * @return number or validation errors; -2 is returned if there are no
      *         sections
      */
-    @edu.umd.cs.findbugs.annotations.SuppressFBWarnings( value="SLF4J_FORMAT_SHOULD_BE_CONST",
-        justification="Error String already evaluated, text from managed class")
-    public int validateAllSections(jmri.util.JmriJFrame frame, LayoutEditor lePanel) {
-        Set<Section> set = getNamedBeanSet();
-        int numSections = 0;
-        int numErrors = 0;
-        if (set.size() <= 0) {
-            return -2;
-        }
-        for (Section section : set) {
-            String s = section.validate(lePanel);
-            if (!s.isEmpty()) {
-                log.error(s);
-                numErrors++;
-            }
-            numSections++;
-        }
-        log.debug("Validated {} Sections - {} errors or warnings.", numSections, numErrors);
-        return numErrors;
-    }
+    public int validateAllSections(jmri.util.JmriJFrame frame, LayoutEditor lePanel);
 
     /**
      * Check direction sensors in SSL for signals.
@@ -184,24 +95,7 @@ public class SectionManager extends AbstractManager<Section> implements Instance
      * @return the number or errors; 0 if no errors; -1 if the panel is null; -2
      *         if there are no sections
      */
-    public int setupDirectionSensors(LayoutEditor lePanel) {
-        if (lePanel == null) {
-            return -1;
-        }
-        Set<Section> set = getNamedBeanSet();
-        int numSections = 0;
-        int numErrors = 0;
-        if (set.size() <= 0) {
-            return -2;
-        }
-        for (Section section : set) {
-            int errors = section.placeDirectionSensors(lePanel);
-            numErrors = numErrors + errors;
-            numSections++;
-        }
-        log.debug("Checked direction sensors for {} Sections - {} errors or warnings.", numSections, numErrors);
-        return numErrors;
-    }
+    public int setupDirectionSensors(LayoutEditor lePanel);
 
     /**
      * Remove direction sensors from SSL for all signals.
@@ -210,159 +104,16 @@ public class SectionManager extends AbstractManager<Section> implements Instance
      * @return the number or errors; 0 if no errors; -1 if the panel is null; -2
      *         if there are no sections
      */
-    public int removeDirectionSensorsFromSSL(LayoutEditor lePanel) {
-        if (lePanel == null) {
-            return -1;
-        }
-        jmri.jmrit.display.layoutEditor.ConnectivityUtil cUtil = lePanel.getConnectivityUtil();
-        Set<Section> set = getNamedBeanSet();
-        if (set.size() <= 0) {
-            return -2;
-        }
-        int numErrors = 0;
-        List<String> sensorList = new ArrayList<>();
-        for (Section s : set) {
-            String name = s.getReverseBlockingSensorName();
-            if ((name != null) && (!name.isEmpty())) {
-                sensorList.add(name);
-            }
-            name = s.getForwardBlockingSensorName();
-            if ((name != null) && (!name.isEmpty())) {
-                sensorList.add(name);
-            }
-        }
-        SignalHeadManager shManager = InstanceManager.getDefault(SignalHeadManager.class);
-        for (SignalHead sh : shManager.getNamedBeanSet()) {
-            if (!cUtil.removeSensorsFromSignalHeadLogic(sensorList, sh)) {
-                numErrors++;
-            }
-        }
-        return numErrors;
-    }
+    public int removeDirectionSensorsFromSSL(LayoutEditor lePanel);
 
     /**
      * Initialize all blocking sensors that exist - set them to 'ACTIVE'.
      */
-    public void initializeBlockingSensors() {
-        for (Section s : getNamedBeanSet()) {
-            try {
-                if (s.getForwardBlockingSensor() != null) {
-                    s.getForwardBlockingSensor().setState(Sensor.ACTIVE);
-                }
-                if (s.getReverseBlockingSensor() != null) {
-                    s.getReverseBlockingSensor().setState(Sensor.ACTIVE);
-                }
-            } catch (JmriException reason) {
-                log.error("Exception when initializing blocking Sensors for Section {}", s.getDisplayName(NamedBean.DisplayOptions.USERNAME_SYSTEMNAME));
-            }
-        }
-    }
-        
+    public void initializeBlockingSensors();
+
     /**
      * Generate Block Sections in stubs/sidings. Called after generating signal logic.
      */
-    
-    
-    public void generateBlockSections() {
-        //find blocks with no paths through i.e. stub (siding)
-        LayoutBlockManager LayoutBlockManager = InstanceManager.getDefault(LayoutBlockManager.class);
-        //print "Layout Block"
-        for (LayoutBlock layoutBlock : LayoutBlockManager.getNamedBeanSet()){
-            if (layoutBlock.getNumberOfThroughPaths() == 0){
-                if (!blockSectionExists(layoutBlock)){
-                    //create block section"
-                    createBlockSection(layoutBlock);
-                }
-            }
-        }
-    }
-                               
-    /**
-     * Check if Block Section already exists
-     * @param layoutBlock
-     * @return true or false
-     */
-    private boolean blockSectionExists(LayoutBlock layoutBlock){
-        
-        for (Section section : getNamedBeanSet()){
-            if (section.getNumBlocks() == 1 
-                    && section.getSectionType() != Section.SIGNALMASTLOGIC 
-                    && layoutBlock.getBlock().equals(section.getEntryBlock())){
-                return true;
-            }
-        }
-        return false;
-    } 
-      
-    private void createBlockSection(LayoutBlock layoutBlock){
-        Section section;
-        try {
-            section = createNewSection(layoutBlock.getUserName());
-        }
-        catch (IllegalArgumentException ex){
-            log.error("Could not create Section from LayoutBlock {}",layoutBlock.getDisplayName());
-            return;
-        }
-        section.addBlock(layoutBlock.getBlock());
-        ArrayList<EntryPoint> entryPointList = new ArrayList<>();
-        Block sb = layoutBlock.getBlock();
-        List <Path> paths = sb.getPaths();
-        for (int j=0; j<paths.size(); j++){
-            Path p = paths.get(j);
-            if (p.getBlock() != sb){
-                //this is path to an outside block, so need an Entry Point
-                String pbDir = Path.decodeDirection(p.getFromBlockDirection());
-                EntryPoint ep = new EntryPoint(sb, p.getBlock(), pbDir);
-                entryPointList.add(ep);
-            }
-        }
-                
-        Block beginBlock = sb;
-        // Set directions where possible
-        List <EntryPoint> epList = getBlockEntryPointsList(beginBlock,entryPointList);
-        if (epList.size() == 1) {
-            (epList.get(0)).setTypeForward();
-        }  
-        Block endBlock = sb;
-        epList = getBlockEntryPointsList(endBlock, entryPointList);
-        if (epList.size() == 1) {
-            (epList.get(0)).setTypeReverse();          
-        }
-            
-        for (int j=0; j<entryPointList.size(); j++){
-            EntryPoint ep = entryPointList.get(j);
-            if (ep.isForwardType()){
-                section.addToForwardList(ep);
-            }else if (ep.isReverseType()){
-                section.addToReverseList(ep);
-            }
-        }
-    }
-            
-    private List <EntryPoint> getBlockEntryPointsList(Block b, List <EntryPoint> entryPointList) {
-        List <EntryPoint> list = new ArrayList<>();
-        for (int i=0; i<entryPointList.size(); i++) {
-            EntryPoint ep = entryPointList.get(i);
-            if (ep.getBlock().equals(b)) {
-                list.add(ep);
-            }
-        }
-        return list;
-    }
-
-    @Override
-    @Nonnull
-    public String getBeanTypeHandled(boolean plural) {
-        return Bundle.getMessage(plural ? "BeanNameSections" : "BeanNameSection");
-    }
-    
-    @Override
-    public void dispose(){
-        InstanceManager.getDefault(SensorManager.class).removeVetoableChangeListener(this);
-        InstanceManager.getDefault(BlockManager.class).removeVetoableChangeListener(this);
-        super.dispose();
-    }
-
-    private final static Logger log = LoggerFactory.getLogger(SectionManager.class);
+    public void generateBlockSections();
 
 }
