+++ java/src/jmri/configurexml/ClassMigration.properties
@@ -71,3 +71,6 @@ jmri.jmrit.display.layoutEditor.configurexml.LayoutXOverXml        = jmri.jmrit.
 jmri.jmrit.display.layoutEditor.configurexml.LayoutRHXOverXml      = jmri.jmrit.display.layoutEditor.configurexml.LayoutRHXOverViewXml
 jmri.jmrit.display.layoutEditor.configurexml.LayoutLHXOverXml      = jmri.jmrit.display.layoutEditor.configurexml.LayoutLHXOverViewXml
 jmri.jmrit.display.layoutEditor.configurexml.LayoutDoubleXOverXml  = jmri.jmrit.display.layoutEditor.configurexml.LayoutDoubleXOverViewXml
+
+# MIgration of Section class in May 2022
+jmri.configurexml.SectionManagerXml         = jmri.managers.configurexml.DefaultSectionManagerXml
