+++ java/test/jmri/SectionManagerTest.java
@@ -13,7 +13,13 @@ public class SectionManagerTest {
 
     @Test
     public void testCTor() {
-        SectionManager t = new SectionManager();
+        SectionManager t = new jmri.managers.DefaultSectionManager();
+        Assert.assertNotNull("exists",t);
+    }
+
+    @Test
+    public void testInstanceManagerAccess() {
+        SectionManager t = InstanceManager.getDefault(SectionManager.class);
         Assert.assertNotNull("exists",t);
     }
