+++ java/test/jmri/SectionTest.java
@@ -2,6 +2,8 @@ package jmri;
 
 import jmri.util.JUnitUtil;
 
+import jmri.implementation.DefaultSection;
+
 import org.junit.jupiter.api.*;
 import org.junit.Assert;
 
@@ -16,17 +18,17 @@ public class SectionTest {
 
    @Test
    public void SysNameConstructorTest(){
-      Assert.assertNotNull("Constructor", new Section("TS1"));
+      Assert.assertNotNull("Constructor", new DefaultSection("TS1"));
    }
 
    @Test
    public void TwoNameStringConstructorTest(){
-      Assert.assertNotNull("Constructor", new Section("TS1", "user name"));
+      Assert.assertNotNull("Constructor", new DefaultSection("TS1", "user name"));
    }
 
    @Test
    public void warnOnBlockAdd() {
-    Section  s = new Section("TS1");
+    Section  s = new DefaultSection("TS1");
     Assert.assertEquals(0, s.getBlockList().size());
     s.addBlock(new Block("IB1", "user"));
     Assert.assertEquals(1, s.getBlockList().size());
@@ -34,13 +36,13 @@ public class SectionTest {
 
    @Test
    public void warnOnBlockAddWithNoUserName() {
-    Section  s = new Section("TS1");
+    Section  s = new DefaultSection("TS1");
     Assert.assertEquals(0, s.getBlockList().size());
     s.addBlock(new Block("IB1"));
     jmri.util.JUnitAppender.assertWarnMessage("Block IB1 does not have a user name, may not work correctly in Section TS1");
     Assert.assertEquals(1, s.getBlockList().size());
    }
-   
+
    @BeforeEach
    public void setUp() {
         JUnitUtil.setUp();
