+++ java/test/jmri/jmrit/dispatcher/AllocatedSectionTest.java
@@ -22,8 +22,8 @@ public class AllocatedSectionTest {
         DispatcherFrame d = InstanceManager.getDefault(DispatcherFrame.class);
         jmri.Transit transit = new jmri.Transit("TT1");
         ActiveTrain at = new ActiveTrain(transit, "Train", ActiveTrain.USER);
-        jmri.Section section1 = new jmri.Section("TS1");
-        jmri.Section section2 = new jmri.Section("TS2");
+        jmri.Section section1 = new jmri.implementation.DefaultSection("TS1");
+        jmri.Section section2 = new jmri.implementation.DefaultSection("TS2");
         AllocatedSection t = new AllocatedSection(section1, at, 1, section2, 2);
         Assert.assertNotNull("exists", t);
         JUnitUtil.dispose(d);
